fix: don't escape HTML in templates
When using `<%= code %>` in `ejs`, it escapes characters by default. This is useful when writing HTML, but in our case, it may sometimes lead to undesirable behavior.

For example, if you generate a project with the description "Satya's cool library", it would escape `'` and output: "Satya&#39 cool library".

This commit swaps the templates to use `<%- code %>` so that nothing is escaped.